pyproject.toml
[project]
name = "bantu"
version = "0.1.0"
description = "Telegram bot for Bantu services"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "python-telegram-bot",
    "python-dotenv",
    "requests",
    "SQLAlchemy",
    "psycopg2-binary",
]
